Accept scale values that are floats besides ints (#184)

+++ src/Document/ContentStream/PositionedText/TextState.php
@@ -8,10 +8,10 @@ use PrinsFrank\PdfParser\Document\Dictionary\DictionaryKey\ExtendedDictionaryKey
 class TextState {
     public function __construct(
         public readonly DictionaryKey|ExtendedDictionaryKey|null $fontName, // Tf
-        public readonly ?float $fontSize,    // Tfs
+        public readonly ?float $fontSize, // Tfs
         public float $charSpace = 0,      // Tc
         public float $wordSpace = 0,      // Tw
-        public int $scale = 100,          // Th
+        public float $scale = 100,        // Th
         public float $leading = 0,        // Tl
         public int $render = 0,           // Tmode
         public float $rise = 0,           // Trise
